ADD function for operator and equals buttons
The operator buttons now updated the operator to the last clicked
operator. The equals button now displays the solution of the equation.
However the behavior for selecting a second number has not been added.
At this point clicking a number after selecting an operator will not
reset the display value.